Split code generation from Xtext UI.
Added an additional feature to be able to install each independently.
