+++ com.cohesionforce.avro.base/pom.xml
@@ -1,0 +1,10 @@
+<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+  <modelVersion>4.0.0</modelVersion>
+	<parent>
+		<version>1.0.0-SNAPSHOT</version>
+		<groupId>com.cohesionforce.avro</groupId>
+		<artifactId>com.cohesionforce.avro.parent</artifactId>
+	</parent>
+	<artifactId>com.cohesionforce.avro.base</artifactId>
+	<packaging>eclipse-plugin</packaging>
+</project>
