+++ com.cohesionforce.avro.base/src/com/cohesionforce/avro/ecore/AvroResource.java
@@ -18,8 +18,6 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
-import com.cohesionforce.avro.AvroEMFConverter;
-
 import org.apache.avro.Schema;
 import org.apache.avro.file.CodecFactory;
 import org.apache.avro.file.DataFileStream;
