+++ com.cohesionforce.avro.gen.feature/build.properties
@@ -1,0 +1,1 @@
+bin.includes = feature.xml
