+++ com.cohesionforce.avro.gen/META-INF/MANIFEST.MF
@@ -1,0 +1,27 @@
+Manifest-Version: 1.0
+Bundle-ManifestVersion: 2
+Bundle-Name: Gen
+Bundle-SymbolicName: com.cohesionforce.avro.gen;singleton:=true
+Bundle-Version: 1.0.0.qualifier
+Bundle-Vendor: COHESIONFORCE
+Automatic-Module-Name: com.cohesionforce.avro.gen
+Bundle-RequiredExecutionEnvironment: JavaSE-1.8
+Require-Bundle: org.eclipse.core.runtime;bundle-version="3.16.0",
+ org.eclipse.core.commands;bundle-version="3.9.500",
+ org.eclipse.core.resources;bundle-version="3.13.500",
+ org.eclipse.emf.ecore;bundle-version="2.19.0",
+ org.eclipse.e4.core.commands;bundle-version="0.12.700",
+ org.eclipse.e4.core.contexts;bundle-version="1.8.200",
+ org.eclipse.e4.core.di.annotations;bundle-version="1.6.400",
+ org.eclipse.e4.core.di;bundle-version="1.7.400",
+ org.eclipse.ui;bundle-version="3.114.0",
+ org.eclipse.e4.ui.services;bundle-version="1.3.600",
+ org.eclipse.emf.codegen.ecore;bundle-version="2.19.0",
+ com.google.guava,
+ org.eclipse.xtext.xbase.lib,
+ org.eclipse.xtend.lib,
+ org.eclipse.xtend.lib.macro,
+ com.cohesionforce.apache.avro;bundle-version="1.7.6",
+ org.eclipse.ui.ide;bundle-version="3.16.0"
+Import-Package: javax.inject;version="1.0.0",
+ org.slf4j
