+++ com.cohesionforce.avro.gen/build.properties
@@ -1,0 +1,6 @@
+source.. = src/,\
+           xtend-gen/
+output.. = bin/
+bin.includes = META-INF/,\
+               .,\
+               plugin.xml
