+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/AvroSchemaGenerator.xtend
