+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/GenerateResourceFactory.xtend
