+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/Utility.java
