+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/handler/GenerateFromGenModelDIHandler.java
@@ -9,8 +9,8 @@ import org.eclipse.e4.core.di.annotations.CanExecute;
 import org.eclipse.e4.core.di.annotations.Execute;
 import org.eclipse.ui.PlatformUI;
 
-import com.google.inject.Inject;
-import com.google.inject.Provider;
+//import com.google.inject.Inject;
+//import com.google.inject.Provider;
 
 public class GenerateFromGenModelDIHandler extends AbstractHandler {
