+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/handler/GenerateJavaDIHandler.java
