+++ com.cohesionforce.avro.gen/src/com/cohesionforce/avro/gen/handler/GenerateJavaHandler.java
