+++ com.cohesionforce.avro.gen/xtend-gen/com/cohesionforce/avro/gen/GenerateResourceFactory.java
