+++ com.cohesionforce.avro.repository/category.xml
@@ -1,11 +1,17 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <site>
-	<feature id="com.cohesionforce.avro.feature" version="0.0.0">
-		<category name="main"/>
-	</feature>
-	<feature id="com.cohesionforce.avro.feature.source" version="0.0.0">
-		<category name="main.source"/>
-	</feature>
-	<category-def name="main" label="AvroSchema"/>
-	<category-def name="main.source" label="AvroSchema (Sources)"/>
+   <feature id="com.cohesionforce.avro.feature" version="0.0.0">
+      <category name="main"/>
+   </feature>
+   <feature id="com.cohesionforce.avro.feature.source" version="0.0.0">
+      <category name="main.source"/>
+   </feature>
+   <feature id="com.cohesionforce.avro.gen.feature">
+      <category name="main"/>
+   </feature>
+   <feature id="com.cohesionforce.avro.gen.feature.source">
+      <category name="main"/>
+   </feature>
+   <category-def name="main" label="AvroSchema"/>
+   <category-def name="main.source" label="AvroSchema (Sources)"/>
 </site>
