+++ com.cohesionforce.avro/META-INF/MANIFEST.MF
@@ -25,15 +25,12 @@ Require-Bundle: org.eclipse.xtext,
  org.eclipse.core.resources;bundle-version="3.13.500",
  org.eclipse.e4.ui.services;bundle-version="1.3.600",
  org.eclipse.xtext.builder,
- org.eclipse.emf.codegen.ecore;bundle-version="2.19.0",
- com.cohesionforce.apache.avro;bundle-version="1.7.6"
+ org.eclipse.emf.codegen.ecore;bundle-version="2.19.0"
 Bundle-RequiredExecutionEnvironment: JavaSE-1.8
 Export-Package: com.cohesionforce.avro,
  com.cohesionforce.avro.avroSchema,
  com.cohesionforce.avro.avroSchema.impl,
  com.cohesionforce.avro.avroSchema.util,
- com.cohesionforce.avro.ecore,
- com.cohesionforce.avro.gen,
  com.cohesionforce.avro.parser.antlr,
  com.cohesionforce.avro.parser.antlr.internal,
  com.cohesionforce.avro.scoping,
